test(manual): add vitest runner and migrate manual-specific tests
`packages/manual` now has its own standalone vitest setup
(`vitest.config.ts` + `tsconfig.json`, no vite dep), so the cross-SSOT
character-equal guard tests (manual yaml `symbols.<id>.description`
matching `shared/symbols.ts`) and the shipped-yaml schema validation
tests now live in `packages/manual/data-validation.test.ts` alongside
the data they exercise. The game-package `manual-schema.test.ts` is
trimmed back to schema-unit only (`collectReferencedSymbols` +
`validateManualSymbols`) and no longer reaches across the workspace
to read manual yamls.

Root `pnpm test` / `pnpm test:run` switched from `--filter game` to
`-r run`, so both packages run their tests in CI. The CI workflow
itself is unchanged because it dispatches via the root script.

Verified: `pnpm -r run test:run` → manual 5/5, game 100/100;
`pnpm build` and `pnpm lint` green; `shared/` untouched. The
pnpm-lock.yaml diff is mostly pnpm 10.33.0 cosmetic reformatting;
the only semantic addition is `vitest@^4.1.5` under
`importers.packages/manual`.

Followup of unify-symbol-description-ssot (PR #54).
