+++ packages/manual/package.json
@@ -3,7 +3,9 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "build": "tsx build.ts"
+    "build": "tsx build.ts",
+    "test": "vitest",
+    "test:run": "vitest run"
   },
   "dependencies": {
     "js-yaml": "^4.1.0"
@@ -11,6 +13,7 @@
   "devDependencies": {
     "@types/js-yaml": "^4.0.0",
     "@types/node": "^25.6.2",
-    "tsx": "^4.0.0"
+    "tsx": "^4.0.0",
+    "vitest": "^4.1.5"
   }
 }
