+++ packages/manual/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "lib": ["ES2020"],
+    "paths": {
+      "@shared/*": ["../../shared/*"]
+    },
+    "types": ["node"]
+  },
+  "include": ["build.ts", "vitest.config.ts", "*.ts", "../../shared"]
+}
